Split node classes and code into node.py, and fix imports
Works towards #130.
