+++ privcount/data_collector.py
@@ -14,8 +14,9 @@ from twisted.internet.protocol import ReconnectingClientFactory
 
 from privcount.counter import SecureCounters, counter_modulus, add_counter_limits_to_config, combine_counters
 from privcount.log import log_error, format_delay_time_wait, format_last_event_time_since
+from privcount.node import PrivCountClient
 from privcount.protocol import PrivCountClientProtocol, TorControlClientProtocol
-from privcount.util import get_public_digest_string, load_public_key_string, encrypt, normalise_path, choose_secret_handshake_path, PrivCountClient
+from privcount.util import get_public_digest_string, load_public_key_string, encrypt, normalise_path, choose_secret_handshake_path
 
 # using reactor: pylint: disable=E1101
 # method docstring missing: pylint: disable=C0111
