+++ privcount/share_keeper.py
@@ -16,7 +16,8 @@ from twisted.internet.protocol import ReconnectingClientFactory
 from privcount.counter import SecureCounters, counter_modulus, add_counter_limits_to_config, combine_counters
 from privcount.log import log_error
 from privcount.protocol import PrivCountClientProtocol
-from util import get_public_digest, generate_keypair, get_serialized_public_key, load_private_key_file, decrypt, normalise_path, choose_secret_handshake_path, PrivCountClient
+from privcount.node import PrivCountClient
+from privcount.util import get_public_digest, generate_keypair, get_serialized_public_key, load_private_key_file, decrypt, normalise_path, choose_secret_handshake_path
 
 class ShareKeeper(ReconnectingClientFactory, PrivCountClient):
     '''
