+++ privcount/tally_server.py
@@ -18,13 +18,13 @@ from twisted.internet.protocol import ServerFactory
 
 from privcount.counter import SecureCounters, counter_modulus, min_blinded_counter_value, max_blinded_counter_value, min_tally_counter_value, max_tally_counter_value, add_counter_limits_to_config, check_noise_weight_config, check_counters_config, CollectionDelay, float_accuracy
 from privcount.log import log_error, format_elapsed_time_since, format_elapsed_time_wait, format_delay_time_until, format_interval_time_between, format_last_event_time_since
+from privcount.node import PrivCountServer, continue_collecting, log_tally_server_status
 from privcount.protocol import PrivCountServerProtocol
 from privcount.statistics_noise import get_noise_allocation
 
 # split lines to help with diffs and readability
-from util import generate_keypair, generate_cert
-from util import normalise_path, choose_secret_handshake_path
-from util import PrivCountServer, continue_collecting, log_tally_server_status
+from privcount.util import generate_keypair, generate_cert
+from privcount.util import normalise_path, choose_secret_handshake_path
 
 # for warning about logging function and format # pylint: disable=W1202
 # for calling methods on reactor # pylint: disable=E1101
